requirements.txt
gradio==5.42.0
pandas==2.2.2
numpy==2.0.2
matplotlib==3.10.0
pillow==11.3.0
scikit-learn==1.6.1
joblib==1.5.1
langchain==0.3.27
langgraph==0.6.5